feature add
    -- select moved axis on button or hat in device tab
    -- auto scroll to last changed axis/button/hat in device tab

+++ joystick_gremlin.py
@@ -399,8 +399,8 @@ class GremlinUi(QtWidgets.QMainWindow):
             gremlin.ui.input_viewer.InputViewerUi()
         geom = self.geometry()
         self.modal_windows["input_viewer"].setGeometry(
-            geom.x() + geom.width() / 2 - 350,
-            geom.y() + geom.height() / 2 - 150,
+            int(geom.x() + geom.width() / 2 - 350),
+            int(geom.y() + geom.height() / 2 - 150),
             700,
             300
         )
